+++ custom_components/mydolphin_plus/number.py
@@ -1,17 +1,18 @@
+from abc import ABC
 import logging
 
-from homeassistant.components.sensor import SensorEntity, SensorEntityDescription
+from homeassistant.components.number import NumberEntity, NumberEntityDescription
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import ATTR_ICON, ATTR_STATE, Platform
+from homeassistant.const import ATTR_STATE, Platform
 from homeassistant.core import HomeAssistant
 
 from .common.base_entity import MyDolphinPlusBaseEntity, async_setup_entities
-from .common.consts import ATTR_ATTRIBUTES
+from .common.consts import ACTION_ENTITY_SET_NATIVE_VALUE, ATTR_ATTRIBUTES
 from .managers.coordinator import MyDolphinPlusCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
-CURRENT_DOMAIN = Platform.SENSOR
+CURRENT_DOMAIN = Platform.NUMBER
 
 
 async def async_setup_entry(
@@ -21,36 +22,37 @@ async def async_setup_entry(
         hass,
         entry,
         CURRENT_DOMAIN,
-        SensorEntityDescription,
-        MyDolphinPlusSensorEntity,
+        NumberEntityDescription,
+        MyDolphinPlusNumberEntity,
         async_add_entities,
     )
 
 
-class MyDolphinPlusSensorEntity(MyDolphinPlusBaseEntity, SensorEntity):
+class MyDolphinPlusNumberEntity(MyDolphinPlusBaseEntity, NumberEntity, ABC):
     """Representation of a sensor."""
 
     def __init__(
         self,
-        entity_description: SensorEntityDescription,
+        entity_description: NumberEntityDescription,
         coordinator: MyDolphinPlusCoordinator,
     ):
         super().__init__(entity_description, coordinator, CURRENT_DOMAIN)
 
-        self._attr_device_class = entity_description.device_class
+        self._attr_native_min_value = entity_description.native_min_value
+        self._attr_native_max_value = entity_description.native_max_value
+
+    async def async_set_native_value(self, value: float) -> None:
+        """Change the selected option."""
+        await self.async_execute_device_action(ACTION_ENTITY_SET_NATIVE_VALUE, value)
 
     def update_component(self, data):
         """Fetch new state parameters for the sensor."""
         if data is not None:
             state = data.get(ATTR_STATE)
             attributes = data.get(ATTR_ATTRIBUTES)
-            icon = data.get(ATTR_ICON)
 
             self._attr_native_value = state
             self._attr_extra_state_attributes = attributes
-
-            if icon is not None:
-                self._attr_icon = icon
 
         else:
             self._attr_native_value = None